move common code to parent BgpAttributesRoute
- Too much code relies on special semantics of `instanceof BgpRoute`
  to be able to safely inherit from `BgpRoute`
  - instead, move needed code into BgpAttributesRoute, from which
    both GeneratedRoute and BgpRoute now inherit
