+++ projects/batfish-common-protocol/src/main/java/org/batfish/datamodel/routing_policy/communities/SetCommunities.java
@@ -7,7 +7,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.Set;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import org.batfish.datamodel.BgpRoute;
+import org.batfish.datamodel.BgpAttributesRoute;
 import org.batfish.datamodel.bgp.community.Community;
 import org.batfish.datamodel.routing_policy.Environment;
 import org.batfish.datamodel.routing_policy.Result;
@@ -28,10 +28,14 @@ public final class SetCommunities extends Statement {
 
   @Override
   public @Nonnull Result execute(Environment environment) {
+    if (!(environment.getOutputRoute() instanceof BgpAttributesRoute.Builder)) {
+      return new Result();
+    }
     CommunityContext ctx = CommunityContext.fromEnvironment(environment);
     CommunitySet communitySet = _communitySetExpr.accept(CommunitySetExprEvaluator.instance(), ctx);
 
-    BgpRoute.Builder<?, ?> bgpRoute = (BgpRoute.Builder<?, ?>) environment.getOutputRoute();
+    BgpAttributesRoute.Builder<?, ?> bgpRoute =
+        (BgpAttributesRoute.Builder<?, ?>) environment.getOutputRoute();
     Set<Community> communities = communitySet.getCommunities();
     bgpRoute.setCommunities(communities);
     if (environment.getWriteToIntermediateBgpAttributes()) {
