+++ projects/batfish-common-protocol/src/test/java/org/batfish/datamodel/matchers/BgpRouteMatchers.java
@@ -6,6 +6,7 @@ import java.util.Set;
 import javax.annotation.Nonnull;
 import javax.annotation.ParametersAreNonnullByDefault;
 import org.batfish.datamodel.AbstractRoute;
+import org.batfish.datamodel.BgpAttributesRoute;
 import org.batfish.datamodel.BgpRoute;
 import org.batfish.datamodel.Bgpv4Route;
 import org.batfish.datamodel.EvpnType3Route;
@@ -27,7 +28,7 @@ public final class BgpRouteMatchers {
    * Provides a matcher that matches when the supplied {@code subMatcher} matches the {@link
    * BgpRoute}'s communities.
    */
-  public static @Nonnull Matcher<BgpRoute<?, ?>> hasCommunities(
+  public static @Nonnull Matcher<BgpAttributesRoute<?, ?>> hasCommunities(
       Matcher<? super Set<Community>> subMatcher) {
     return new HasCommunities(subMatcher);
   }
@@ -36,7 +37,8 @@ public final class BgpRouteMatchers {
    * Provides a matcher that matches when the supplied {@code expectedOriginType} is equal to the
    * {@link BgpRoute}'s originType.
    */
-  public static @Nonnull Matcher<BgpRoute<?, ?>> hasOriginType(OriginType expectedOriginType) {
+  public static @Nonnull Matcher<BgpAttributesRoute<?, ?>> hasOriginType(
+      OriginType expectedOriginType) {
     return new HasOriginType(equalTo(expectedOriginType));
   }
 
@@ -44,7 +46,7 @@ public final class BgpRouteMatchers {
    * Provides a matcher that matches when the supplied {@code expectedWeight} is equal to the {@link
    * BgpRoute}'s weight.
    */
-  public static @Nonnull Matcher<BgpRoute<?, ?>> hasWeight(int expectedWeight) {
+  public static @Nonnull Matcher<BgpAttributesRoute<?, ?>> hasWeight(int expectedWeight) {
     return new HasWeight(equalTo(expectedWeight));
   }
